refactor: creates stubs file and constants
