--- src/config/index.ts
@@ -1,1 +1,0 @@
-export const CREDENTIALS_KEY = 'account.credentials';
