+++ src/utils/constants/index.ts
@@ -1,0 +1,4 @@
+export const CREDENTIALS_KEY = 'account.credentials';
+export const hexadecimalRegex = /^[a-f0-9]+$/;
+export const expirationWhenRememberIsOnn = 36e5 * 24 * 7;
+export const expirationWhenRememberIsOff = 36e5;
